extract menuicon css into separate file
